move execution-engine to a separate dir
quite initial, not chaniging interfaces too much
