+++ arbnode/execution/arb_interface.go
@@ -1,7 +1,7 @@
 // Copyright 2021-2022, Offchain Labs, Inc.
 // For license information, see https://github.com/nitro/blob/master/LICENSE
 
-package arbnode
+package execution
 
 import (
 	"context"
@@ -20,32 +20,28 @@ type TransactionPublisher interface {
 }
 
 type ArbInterface struct {
-	txStreamer  *TransactionStreamer
+	exec        *ExecutionEngine
 	txPublisher TransactionPublisher
-	arbNode     *Node
+	arbNode     interface{}
 }
 
-func NewArbInterface(txStreamer *TransactionStreamer, txPublisher TransactionPublisher) (*ArbInterface, error) {
+func NewArbInterface(exec *ExecutionEngine, txPublisher TransactionPublisher) (*ArbInterface, error) {
 	return &ArbInterface{
-		txStreamer:  txStreamer,
+		exec:        exec,
 		txPublisher: txPublisher,
 	}, nil
 }
 
-func (a *ArbInterface) Initialize(n *Node) {
-	a.arbNode = n
+func (a *ArbInterface) Initialize(arbnode interface{}) {
+	a.arbNode = arbnode
 }
 
 func (a *ArbInterface) PublishTransaction(ctx context.Context, tx *types.Transaction) error {
 	return a.txPublisher.PublishTransaction(ctx, tx)
 }
 
-func (a *ArbInterface) TransactionStreamer() *TransactionStreamer {
-	return a.txStreamer
-}
-
 func (a *ArbInterface) BlockChain() *core.BlockChain {
-	return a.txStreamer.exec.bc
+	return a.exec.bc
 }
 
 func (a *ArbInterface) ArbNode() interface{} {
