+++ arbnode/sync_monitor.go
@@ -64,9 +64,7 @@ func (s *SyncMonitor) SyncProgressMap() map[string]interface{} {
 	}
 	res["broadcasterQueuedMessagesPos"] = broadcasterQueuedMessagesPos
 
-	lastBlockNum := s.txStreamer.exec.bc.CurrentHeader().Number.Uint64()
-	res["blockNum"] = lastBlockNum
-	lastBuiltMessage, err := s.txStreamer.exec.BlockNumberToMessageCount(lastBlockNum)
+	lastBuiltMessage, err := s.txStreamer.exec.HeadMessageNumber()
 	if err != nil {
 		res["blockMessageToMessageCountError"] = err.Error()
 		syncing = true
